+++ packages/moonstone/Scroller/Scroller.js
@@ -194,6 +194,13 @@ class ScrollerBase extends Component {
 				// set scroll position so that the top of the container is at least on the top as a fallback.
 				newScrollTop = newItemTop - nestedItemHeight;
 			}
+
+			// Ensure that the adjusted scrollTop would at least scroll the container to the top of
+			// the viewport (e.g. because the container is at the bottom of the scroller and the
+			// nested item is at the top of the container)
+			if (newItemTop > newScrollTop) {
+				newScrollTop = newItemTop;
+			}
 		} else if (itemBottom - scrollBottom > epsilon) {
 			// Caculate when 5-way focus down past the bottom.
 			newScrollTop += itemBottom - scrollBottom;
@@ -429,6 +436,7 @@ Scroller.defaultProps = {
 
 /**
  * A Moonstone-styled native Scroller, Scrollable applied.
+ *
  * For smooth native scrolling, web engine with below Chromium 61, should be launched
  * with the flag '--enable-blink-features=CSSOMSmoothScroll' to support it.
  * The one with Chromium 61 or above, is launched to support it by default.