+++ packages/moonstone/Skinnable/Skinnable.js
@@ -6,7 +6,7 @@
  */
 
 import hoc from '@enact/core/hoc';
-import SkinnableBase, {withSkinnableProps} from '@enact/ui/Skinnable';
+import SkinnableBase from '@enact/ui/Skinnable';
 
 const defaultConfig = {
 	skins: {
@@ -16,10 +16,11 @@ const defaultConfig = {
 };
 
 /**
- * This higher-order component is based on [ui/Skinnable]{@link ui/Skinnable.Skinnable} and comes
- * pre-configured for Moonstone's supported skins: "dark" (default) and "light". It is used to apply
- * the relevant skinning classes to each component and has been used to pre-select specific skins
- * for some components.
+ * This higher-order component is based on [ui/Skinnable]{@link ui/Skinnable.Skinnable}.
+ *
+ * `Skinnable` comes pre-configured for Moonstone's supported skins: "dark" (default) and "light".
+ * It is used to apply the relevant skinning classes to each component and has been used to
+ * pre-select specific skins for some components.
  *
  * @class Skinnable
  * @memberof moonstone/Skinnable
@@ -29,10 +30,11 @@ const defaultConfig = {
 const Skinnable = hoc(defaultConfig, SkinnableBase);
 
 /**
- * Select a skin by name by specifying this property. Available Moonstone skins are
- * `"dark"` (default) and `"light"`. This may be changed at runtime. All components already use
- * their defaults, but a skin may be changed via this prop or by using {@link moonstone/Skinnable}
- * directly and a config object.
+ * Select a skin by name by specifying this property.
+ *
+ * Available Moonstone skins are `"dark"` (default) and `"light"`. This may be changed at runtime.
+ * All components already use their defaults, but a skin may be changed via this prop or by using
+ * `Skinnable` directly and a config object.
  *
  * Example:
  * ```
@@ -49,6 +51,5 @@ const Skinnable = hoc(defaultConfig, SkinnableBase);
 
 export default Skinnable;
 export {
-	Skinnable,
-	withSkinnableProps
+	Skinnable
 };