+++ packages/moonstone/ToggleIcon/ToggleIcon.js
@@ -1,11 +1,13 @@
 /**
- * Provides Moonstone-themed Icon component with interactive togglable capabilities, upon which many
- * moonstone Icon components are based. Ex:
- * [Checkbox]{@link moonstone/Checkbox.Checkbox},
- * [FormCheckbox]{@link moonstone/FormCheckbox.FormCheckbox},
- * [Switch]{@link moonstone/Switch.Switch},
- * [the icon for RadioItem]{@link moonstone/RadioItem.RadioItem}, and
- * [the icon for SelectableItem]{@link moonstone/SelectableItem.SelectableItem}.
+ * Provides Moonstone-themed Icon component with interactive togglable capabilities.
+ *
+ * Moonstone components that uses `ToggleIcon`:
+ *
+ * * [Checkbox]{@link moonstone/Checkbox.Checkbox},
+ * * [FormCheckbox]{@link moonstone/FormCheckbox.FormCheckbox},
+ * * [Switch]{@link moonstone/Switch.Switch},
+ * * [RadioItem]{@link moonstone/RadioItem.RadioItem}, and
+ * * [SelectableItem]{@link moonstone/SelectableItem.SelectableItem}.
  *
  * @module moonstone/ToggleIcon
  * @example
@@ -47,7 +49,7 @@ const ToggleIconBase = kind({
 });
 
 /**
- * Moonstone-specific behaviors to apply to [ToggleIconBase]{@link moonstone/ToggleIcon.ToggleIconBase}.
+ * Moonstone-specific behaviors to apply to `ToggleIconBase`.
  *
  * @hoc
  * @memberof moonstone/ToggleIcon