+++ packages/moonstone/TooltipDecorator/TooltipDecorator.js
@@ -26,10 +26,11 @@ let currentTooltip; // needed to know whether or not we should stop a showing jo
  */
 const defaultConfig = {
 	/**
-	 * The name of the property which will receive the tooltip node. By default, `TooltipDecorator`
-	 * will add a new child to the wrapped component, following any other children passed in. If
-	 * a component needs to, it can specify another property to receive the tooltip and the
-	 * `children` property will not be modified.
+	 * The name of the property which will receive the tooltip node.
+	 *
+	 * By default, `TooltipDecorator` will add a new child to the wrapped component, following any
+	 * other children passed in. If a component needs to, it can specify another property to receive
+	 * the tooltip and the `children` property will not be modified.
 	 *
 	 * @type {String}
 	 * @default 'children'
@@ -168,6 +169,12 @@ const TooltipDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			};
 		}
 
+		componentDidUpdate (prevProps) {
+			if (this.state.showing && prevProps.tooltipText !== this.props.tooltipText) {
+				this.setTooltipLayout();
+			}
+		}
+
 		componentWillUnmount () {
 			if (currentTooltip === this) {
 				currentTooltip = null;
@@ -180,6 +187,8 @@ const TooltipDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		setTooltipLayout () {
+			if (!this.tooltipRef || !this.clientRef) return;
+
 			const position = this.props.tooltipPosition;
 			const arr = position.split(' ');
 			let tooltipDirection = null;