+++ packages/sampler/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The following is a curated list of changes in the Enact Sampler, newest changes on the top.
 
+## [2.1.2] - 2018-09-04
+
+No significant changes.
+
+## [2.1.1] - 2018-08-27
+
+No significant changes.
+
 ## [2.1.0] - 2018-08-20
 
 ### Fixed