+++ packages/ui/CHANGELOG.md
@@ -4,10 +4,20 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ## [unreleased]
 
-### Changed
+### Fixed
 
-- `ui/Icon` to require `children`
-- `ui/LabeledIcon` to require `icon` prop
+- `ui/Marquee` to stop when blurred during restart timer
+
+## [2.1.2] - 2018-09-04
+
+### Fixed
+
+- `ui/GridListImageItem` to properly set `selected` style
+- `ui/Marquee` positioning bug when used with CSS flexbox layouts
+
+## [2.1.1] - 2018-08-27
+
+No significant changes.
 
 ## [2.1.0] - 2018-08-20